requirements.txt
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==2.8.0
exceptiongroup==1.3.1
fonttools==4.60.1
idna==3.10
iniconfig==2.3.0
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
nltk==3.9.2
numpy==2.2.6
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
pybind11==3.0.1
Pygments==2.19.2
pymongo==4.15.1
pyparsing==3.2.5
pytest==9.0.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
six==1.17.0
soupsieve==2.8
tomli==2.3.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
